refactor: move component in templates folder
