+++ packages/apps/stock/src/components/index.js
@@ -18,3 +18,4 @@
 
 export * from './molecules';
 export * from './organisms';
+export * from './templates';
