+++ packages/apps/stock/src/components/organisms/index.js
@@ -16,47 +16,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export {default as CardStockIndicator} from './CardStockIndicator/CardStockIndicator';
-export {default as CustomerDeliveryCard} from './CustomerDeliveryCard/CustomerDeliveryCard';
-export {default as CustomerDeliveryLineCard} from './CustomerDeliveryLineCard/CustomerDeliveryLineCard';
-export {default as CustomerDeliveryLineDetailFixedItems} from './CustomerDeliveryLineDetailFixedItems/CustomerDeliveryLineDetailFixedItems';
-export {default as CustomerDeliveryLineQuantityCard} from './CustomerDeliveryLineQuantityCard/CustomerDeliveryLineQuantityCard';
-export {default as CustomerDeliveryDetailHeader} from './CustomerDeliveryDetailHeader/CustomerDeliveryDetailHeader';
 export {default as DescriptionCard} from './DescriptionCard/DescriptionCard';
-export {default as InternalMoveCard} from './InternalMoveCard/InternalMoveCard';
-export {default as InternalMoveGeneralBody} from './InternalMoveGeneralBody/InternalMoveGeneralBody';
-export {default as InternalMoveGeneralFixedItems} from './InternalMoveGeneralFixedItems/InternalMoveGeneralFixedItems';
-export {default as InternalMoveLineCard} from './InternalMoveLineCard/InternalMoveLineCard';
-export {default as InternalMoveLineDetailsFixedItems} from './InternalMoveLineDetailsFixedItems/InternalMoveLineDetailsFixedItems';
-export {default as InternalMoveLineNotes} from './InternalMoveLineNotes/InternalMoveLineNotes';
-export {default as InventoryCard} from './InventoryCard/InventoryCard';
-export {default as InventoryHeader} from './InventoryHeader/InventoryHeader';
-export {default as InventoryLineCard} from './InventoryLineCard/InventoryLineCard';
-export {default as InventoryLineDetailsQuantityCard} from './InventoryLineDetailsQuantityCard/InventoryLineDetailsQuantityCard';
-export {default as InventoryPlannedDetailsHeader} from './InventoryPlannedDetailsHeader/InventoryPlannedDetailsHeader';
-export {default as InventoryStartedFixedItems} from './InventoryStartedFixedItems/InventoryStartedFixedItems';
-export {default as InventoryStartedHeader} from './InventoryStartedHeader/InventoryStartedHeader';
-export {default as ProductCard} from './ProductCard/ProductCard';
-export {default as ProductCardInfo} from './ProductCardInfo/ProductCardInfo';
-export {default as ProductCardStockIndicatorList} from './ProductCardStockIndicatorList/ProductCardStockIndicatorList';
-export {default as ProductCharacteristics} from './ProductCharacteristics/ProductCharacteristics';
-export {default as ProductFixedItems} from './ProductFixedItems/ProductFixedItems';
-export {default as ProductPacking} from './ProductPacking/ProductPacking';
-export {default as ProductSeeDistributionStocksLocation} from './ProductSeeDistributionStocksLocation/ProductSeeDistributionStocksLocation';
-export {default as ProductSmallPropertyCardList} from './ProductSmallPropertyCardList/ProductSmallPropertyCardList';
-export {default as ProductStockHeader} from './ProductStockHeader/ProductStockHeader';
-export {default as ProductStockLocationCard} from './ProductStockLocationCard/ProductStockLocationCard';
-export {default as ProductVariantCard} from './ProductVariantCard/ProductVariantCard';
 export {default as QuantityCard} from './QuantityCard/QuantityCard';
 export {default as SmallPropertyCard} from './SmallPropertyCard/SmallPropertyCard';
-export {default as StockCorrectionBody} from './StockCorrectionBody/StockCorrectionBody';
-export {default as StockCorrectionFixedItems} from './StockCorrectionFixedItems/StockCorrectionFixedItems';
-export {default as StockCorrectionHeader} from './StockCorrectionHeader/StockCorrectionHeader';
-export {default as StockMoveHeader} from './StockMoveHeader/StockMoveHeader';
-export {default as SupplierArrivalAddTracking} from './SupplierArrivalAddTracking/SupplierArrivalAddTracking';
-export {default as SupplierArrivalCard} from './SupplierArrivalCard/SupplierArrivalCard';
-export {default as SupplierArrivalDetailsHeader} from './SupplierArrivalDetailsHeader/SupplierArrivalDetailsHeader';
-export {default as SupplierArrivalLineCard} from './SupplierArrivalLineCard/SupplierArrivalLineCard';
-export {default as SupplierArrivalLineDetailFixedItems} from './SupplierArrivalLineDetailFixedItems/SupplierArrivalLineDetailFixedItems';
-export {default as SupplierArrivalLineDetailQuantityCard} from './SupplierArrivalLineDetailQuantityCard/SupplierArrivalLineDetailQuantityCard';
-export {default as SupplierProductInfo} from './SupplierProductInfo/SupplierProductInfo';
