+++ packages/apps/stock/src/components/templates/CustomerDeliveryCard/CustomerDeliveryCard.tsx
