+++ packages/apps/stock/src/components/templates/CustomerDeliveryDetailHeader/CustomerDeliveryDetailHeader.js
