+++ packages/apps/stock/src/components/templates/CustomerDeliveryLineDetailFixedItems/CustomerDeliveryLineDetailFixedItems.js
