+++ packages/apps/stock/src/components/templates/InternalMoveGeneralBody/InternalMoveGeneralBody.js
