+++ packages/apps/stock/src/components/templates/InternalMoveGeneralFixedItems/InternalMoveGeneralFixedItems.js
