+++ packages/apps/stock/src/components/templates/InternalMoveLineCard/InternalMoveLineCard.tsx
