+++ packages/apps/stock/src/components/templates/InternalMoveLineDetailsFixedItems/InternalMoveLineDetailsFixedItems.js
