+++ packages/apps/stock/src/components/templates/InventoryCard/InventoryCard.tsx
