+++ packages/apps/stock/src/components/templates/InventoryHeader/InventoryHeader.tsx
