+++ packages/apps/stock/src/components/templates/InventoryLineCard/InventoryLineCard.tsx
