+++ packages/apps/stock/src/components/templates/InventoryPlannedDetailsHeader/InventoryPlannedDetailsHeader.js
