+++ packages/apps/stock/src/components/templates/InventoryStartedFixedItems/InventoryStartedFixedItems.js
