+++ packages/apps/stock/src/components/templates/InventoryStartedHeader/InventoryStartedHeader.js
