+++ packages/apps/stock/src/components/templates/ProductCard/ProductCard.tsx
