+++ packages/apps/stock/src/components/templates/ProductCardInfo/ProductCardInfo.tsx
