+++ packages/apps/stock/src/components/templates/ProductCharacteristics/ProductCharacteristics.tsx
