+++ packages/apps/stock/src/components/templates/ProductFixedItems/ProductFixedItems.js
