+++ packages/apps/stock/src/components/templates/ProductSeeDistributionStocksLocation/ProductSeeDistributionStocksLocation.js
