+++ packages/apps/stock/src/components/templates/ProductStockHeader/ProductStockHeader.js
