+++ packages/apps/stock/src/components/templates/ProductStockLocationCard/ProductStockLocationCard.tsx
