+++ packages/apps/stock/src/components/templates/ProductVariantCard/ProductVariantCard.tsx
