+++ packages/apps/stock/src/components/templates/StockCorrectionFixedItems/StockCorrectionFixedItems.js
