+++ packages/apps/stock/src/components/templates/StockCorrectionHeader/StockCorrectionHeader.js
