+++ packages/apps/stock/src/components/templates/SupplierArrivalCard/SupplierArrivalCard.tsx
