+++ packages/apps/stock/src/components/templates/SupplierArrivalLineCard/SupplierArrivalLineCard.tsx
