+++ packages/apps/stock/src/components/templates/SupplierArrivalLineDetailFixedItems/SupplierArrivalLineDetailFixedItems.js
